# docs/design/openapi.yaml
openapi: 3.1.0
info:
  title: RMS Implementation Architecture Baseline 2.0 — OpenAPI draft
  version: 2.0.0-draft
  description: Design only; NOT IMPLEMENTED. 业务来源唯一冻结Baseline1.3。服务器根/api/v1。全部保护操作检查session、启用状态、实际角色并集；自我评审等限制独立执行。请求只白名单，历史不可变，无generic
    status API。SQL BIGINT IDs/lockVersion采用十进制字符串。两个/auth技术GET为实现派生，不新增FR。数据库generated活动键不公开。密码原值与1..72 UTF-8 bytes技术政策已冻结；非admin assignee目录待决。
servers:
- url: /api/v1
security:
- cookieSession: []
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags:
      - user
      summary: 管理员分页读取安全资料；不得被非管理员作为用户目录调用
      description: 管理员分页读取安全资料；不得被非管理员作为用户目录调用。仅启用账号；服务端重新检查角色并集及独立业务限制。
      x-module: user
      x-fr:
      - FR-USER-004
      x-br:
      - BR-AUTH-001
      - BR-AUTH-002
      - BR-AUTH-003
      - BR-AUTH-004
      - BR-USER-003
      x-uc:
      - UC-02
      x-ac:
      - AC-01
      x-required-roles: &id002
      - ADMIN
      x-transaction: READ_ONLY
      responses:
        '200':
          description: 成功；具体业务含义见操作说明
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminUserResponsePage'
        '400':
          description: 字段/完整性错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: 未认证/凭证错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: 权限/启用状态/自我评审/CSRF拒绝
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: 目标不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: 修订/状态/唯一/基准/图锁冲突
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '500':
          description: 内部错误；业务写全回滚
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
      parameters:
      - $ref: '#/components/parameters/Page'
      - $ref: '#/components/parameters/Size'
    post:
      operationId: createUser
      tags:
      - user
      summary: 管理员创建用户和至少一预定义角色，服务端BCrypt，不返回密码
      description: 管理员创建用户和至少一预定义角色，服务端BCrypt，不返回密码。仅启用账号；服务端重新检查角色并集及独立业务限制。
      x-module: user
      x-fr:
      - FR-AUDIT-001
      - FR-USER-001
      - FR-USER-005
      x-br:
      - BR-AUDIT-001
      - BR-AUDIT-002
      - BR-AUDIT-003
      - BR-AUDIT-004
      - BR-AUDIT-006
      - BR-AUTH-001
      - BR-AUTH-002
      - BR-AUTH-003
      - BR-AUTH-004
      - BR-USER-001
      - BR-USER-002
      - BR-USER-003
      - BR-USER-004
      - BR-USER-005
      x-uc:
      - UC-02
      - UC-S1
      x-ac:
      - AC-01
      - AC-10
      x-required-roles: *id002
      x-transaction: APPLICATION_REQUIRED
      responses:
        '201':
          description: 成功；具体业务含义见操作说明
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminUserResponse'
        '400':
          description: 字段/完整性错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: 未认证/凭证错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: 权限/启用状态/自我评审/CSRF拒绝
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: 目标不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: 修订/状态/唯一/基准/图锁冲突
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '500':
          description: 内部错误；业务写全回滚
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
      parameters:
      - $ref: '#/components/parameters/CsrfHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
  /users/{id}:
    patch:
      operationId: editUser
      tags:
      - user
      summary: 更新用户安全基本资料，保留标识和历史，不能更新密码
      description: 更新用户安全基本资料，保留标识和历史，不能更新密码。仅启用账号；服务端重新检查角色并集及独立业务限制。
      x-module: user
      x-fr:
      - FR-AUDIT-001
      - FR-USER-002
      x-br:
      - BR-AUDIT-001
      - BR-AUDIT-002
      - BR-AUDIT-003
      - BR-AUDIT-004
      - BR-AUDIT-006
      - BR-AUTH-001
      - BR-AUTH-002
      - BR-AUTH-003
      - BR-AUTH-004
      - BR-USER-001
      - BR-USER-003
      x-uc:
      - UC-02
      - UC-S1
      x-ac:
      - AC-01
      - AC-10
      x-required-roles: *id002
      x-transaction: APPLICATION_REQUIRED
      responses:
        '200':
          description: 成功；具体业务含义见操作说明
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminUserResponse'
        '400':
          description: 字段/完整性错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: 未认证/凭证错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: 权限/启用状态/自我评审/CSRF拒绝
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: 目标不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: 修订/状态/唯一/基准/图锁冲突
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '500':
          description: 内部错误；业务写全回滚
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
      parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/Id'
      - $ref: '#/components/parameters/CsrfHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EditUserRequest'
  /users/{id}/roles:
    put:
      operationId: setUserRoles
      tags:
      - user
      summary: 替换预定义角色集合，启用用户不能无角色，grantedBy来自当前ADMIN
      description: 替换预定义角色集合，启用用户不能无角色，grantedBy来自当前ADMIN。仅启用账号；服务端重新检查角色并集及独立业务限制。
      x-module: user
      x-fr:
      - FR-AUDIT-001
      - FR-USER-005
      - FR-USER-006
      x-br:
      - BR-AUDIT-001
      - BR-AUDIT-002
      - BR-AUDIT-003
      - BR-AUDIT-004
      - BR-AUDIT-006
      - BR-AUTH-001
      - BR-AUTH-002
      - BR-AUTH-003
      - BR-AUTH-004
      - BR-CR-019
      - BR-REV-004
      - BR-USER-002
      - BR-USER-003
      - BR-USER-004
      - BR-USER-005
      x-uc:
      - UC-02
      - UC-S1
      x-ac:
      - AC-01
      - AC-10
      x-required-roles: *id002
      x-transaction: APPLICATION_REQUIRED
      responses:
        '200':
          description: 成功；具体业务含义见操作说明
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminUserResponse'
        '400':
          description: 字段/完整性错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: 未认证/凭证错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: 权限/启用状态/自我评审/CSRF拒绝
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: 目标不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: 修订/状态/唯一/基准/图锁冲突
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '500':
          description: 内部错误；业务写全回滚
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
      parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/Id'
      - $ref: '#/components/parameters/CsrfHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetRolesRequest'
  /users/{id}/enable:
    post:
      operationId: enableUser
      tags:
      - user
      summary: 管理员enable账号；每次保护操作检查实际启用状态
      description: 管理员enable账号；每次保护操作检查实际启用状态。仅启用账号；服务端重新检查角色并集及独立业务限制。
      x-module: user
      x-fr:
      - FR-AUDIT-001
      - FR-USER-003
      x-br:
      - BR-AUDIT-001
      - BR-AUDIT-002
      - BR-AUDIT-003
      - BR-AUDIT-004
      - BR-AUDIT-006
      - BR-AUTH-001
      - BR-AUTH-002
      - BR-AUTH-003
      - BR-AUTH-004
      - BR-USER-003
      x-uc:
      - UC-02
      - UC-S1
      x-ac:
      - AC-01
      - AC-10
      x-required-roles: *id002
      x-transaction: APPLICATION_REQUIRED
      responses:
        '200':
          description: 成功；具体业务含义见操作说明
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminUserResponse'
        '400':
          description: 字段/完整性错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: 未认证/凭证错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: 权限/启用状态/自我评审/CSRF拒绝
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: 目标不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: 修订/状态/唯一/基准/图锁冲突
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '500':
          description: 内部错误；业务写全回滚
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
      parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/Id'
      - $ref: '#/components/parameters/CsrfHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LockCommand'
  /users/{id}/disable:
    post:
      operationId: disableUser
      tags:
      - user
      summary: 管理员disable账号；每次保护操作检查实际启用状态
      description: 管理员disable账号；每次保护操作检查实际启用状态。仅启用账号；服务端重新检查角色并集及独立业务限制。
      x-module: user
      x-fr:
      - FR-AUDIT-001
      - FR-USER-003
      x-br:
      - BR-AUDIT-001
      - BR-AUDIT-002
      - BR-AUDIT-003
      - BR-AUDIT-004
      - BR-AUDIT-006
      - BR-AUTH-001
      - BR-AUTH-002
      - BR-AUTH-003
      - BR-AUTH-004
      - BR-USER-003
      x-uc:
      - UC-02
      - UC-S1
      x-ac:
      - AC-01
      - AC-10
      x-required-roles: *id002
      x-transaction: APPLICATION_REQUIRED
      responses:
        '200':
          description: 成功；具体业务含义见操作说明
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminUserResponse'
        '400':
          description: 字段/完整性错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: 未认证/凭证错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: 权限/启用状态/自我评审/CSRF拒绝
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: 目标不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: 修订/状态/唯一/基准/图锁冲突
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '500':
          description: 内部错误；业务写全回滚
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
      parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/Id'
      - $ref: '#/components/parameters/CsrfHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LockCommand'
...
  /requirements/{id}/versions:
    get:
      operationId: listVersions
      tags:
      - version
      summary: 读取全部正式版本；首次批准前为空集合
      description: 读取全部正式版本；首次批准前为空集合。仅启用账号；服务端重新检查角色并集及独立业务限制。
      x-module: version
      x-fr:
      - FR-VER-002
      - FR-VER-004
      - FR-VER-005
      x-br:
      - BR-AUTH-001
      - BR-AUTH-002
      - BR-AUTH-003
      - BR-AUTH-004
      - BR-VER-001
      - BR-VER-004
      - BR-VER-007
      - BR-VER-008
      - BR-VER-009
      - BR-VER-010
      - BR-VER-011
      x-uc:
      - UC-08
      x-ac:
      - AC-05
      x-required-roles: *id001
      x-transaction: READ_ONLY
      responses:
        '200':
          description: 成功；具体业务含义见操作说明
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RequirementVersionResponsePage'
        '400':
          description: 字段/完整性错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: 未认证/凭证错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: 权限/启用状态/自我评审/CSRF拒绝
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: 目标不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: 修订/状态/唯一/基准/图锁冲突
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '500':
          description: 内部错误；业务写全回滚
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
      parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/Id'
      - $ref: '#/components/parameters/Page'
      - $ref: '#/components/parameters/Size'
  /requirements/{id}/versions/{versionId}:
    get:
      operationId: getVersion
      tags:
      - version
      summary: 按需求和版本读取完整八项只读历史、产生者、原因和唯一来源
      description: 按需求和版本读取完整八项只读历史、产生者、原因和唯一来源。仅启用账号；服务端重新检查角色并集及独立业务限制。
      x-module: version
      x-fr:
      - FR-VER-003
      - FR-VER-004
      x-br:
      - BR-AUTH-001
      - BR-AUTH-002
      - BR-AUTH-003
      - BR-AUTH-004
      - BR-VER-004
      - BR-VER-005
      - BR-VER-007
      - BR-VER-009
      - BR-VER-010
      x-uc:
      - UC-08
      x-ac:
      - AC-05
      x-required-roles: *id001
      x-transaction: READ_ONLY
      responses:
        '200':
          description: 成功；具体业务含义见操作说明
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RequirementVersionResponse'
        '400':
          description: 字段/完整性错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: 未认证/凭证错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: 权限/启用状态/自我评审/CSRF拒绝
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: 目标不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: 修订/状态/唯一/基准/图锁冲突
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '500':
          description: 内部错误；业务写全回滚
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
      parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/Id'
      - name: versionId
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/Id'
...
components:
  securitySchemes:
    cookieSession:
      type: apiKey
      in: cookie
      name: JSESSIONID
      description: HttpOnly Secure SameSite=Lax，同源session；所有unsafe请求另需X-CSRF-TOKEN，包括login/logout。
  parameters:
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 0
        default: 0
    Size:
      name: size
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    CsrfHeader:
      name: X-CSRF-TOKEN
      in: header
      required: true
      schema:
        type: string
        minLength: 1
      description: GET /auth/csrf取得；登录/退出后刷新；不是actor身份。
  schemas:
    Id:
      type: string
      pattern: ^[1-9][0-9]*$
      maxLength: 19
      description: SQL signed BIGINT正整数十进制字符串；服务端再校验<=9223372036854775807，避免JavaScript Number精度丢失。
    LockVersion:
      type: string
      pattern: ^(0|[1-9][0-9]*)$
      maxLength: 19
      description: 原有signed BIGINT lockVersion非负十进制字符串；服务端校验<=9223372036854775807、比较并由@Version更新。
    RoleCode:
      type: string
      enum:
      - ADMIN
      - REQUIREMENT_ENGINEER
      - REVIEWER
      - PROJECT_MEMBER
      - VIEWER
...
    RequirementVersionResponse:
      type: object
      additionalProperties: false
      properties:
        versionId:
          type: string
          pattern: ^[1-9][0-9]*$
          maxLength: 19
          description: SQL signed BIGINT正整数十进制字符串；服务端再校验<=9223372036854775807，避免JavaScript Number精度丢失。
          readOnly: true
        requirementId:
          type: string
          pattern: ^[1-9][0-9]*$
          maxLength: 19
          description: SQL signed BIGINT正整数十进制字符串；服务端再校验<=9223372036854775807，避免JavaScript Number精度丢失。
          readOnly: true
        versionNo:
          type: integer
          minimum: 1
          readOnly: true
        title:
          type: string
          maxLength: 200
          readOnly: true
        description:
          type: string
          readOnly: true
        level:
          type: string
          maxLength: 32
          enum:
          - BUSINESS
          - USER
          - SYSTEM
          readOnly: true
        kind:
          type: string
          maxLength: 32
          enum:
          - FUNCTIONAL
          - QUALITY
          - CONSTRAINT
          readOnly: true
        priority:
          type: string
          maxLength: 32
          enum:
          - LOW
          - MEDIUM
          - HIGH
          - CRITICAL
          readOnly: true
        source:
          type: string
          readOnly: true
        rationale:
          type: string
          readOnly: true
        acceptanceCriteria:
          type: string
          readOnly: true
        initialReviewId:
          type:
          - string
          - 'null'
          pattern: ^[1-9][0-9]*$
          maxLength: 19
          description: SQL signed BIGINT正整数十进制字符串；服务端再校验<=9223372036854775807，避免JavaScript Number精度丢失。
          readOnly: true
        appliedChangeRequestId:
          type:
          - string
          - 'null'
          pattern: ^[1-9][0-9]*$
          maxLength: 19
          description: SQL signed BIGINT正整数十进制字符串；服务端再校验<=9223372036854775807，避免JavaScript Number精度丢失。
          readOnly: true
        createdBy:
          type: string
          pattern: ^[1-9][0-9]*$
          maxLength: 19
          description: SQL signed BIGINT正整数十进制字符串；服务端再校验<=9223372036854775807，避免JavaScript Number精度丢失。
          readOnly: true
        createdAt:
          type: string
          format: date-time
          pattern: .*Z$
          description: 服务端UTC，ISO8601 Z；SQL DATETIME(6)，微秒精度。
          readOnly: true
        changeReason:
          type: string
          readOnly: true
      required:
      - versionId
      - requirementId
      - versionNo
      - title
      - description
      - level
      - kind
      - priority
      - source
      - rationale
      - acceptanceCriteria
      - initialReviewId
      - appliedChangeRequestId
      - createdBy
      - createdAt
      - changeReason
      description: 安全传输投影；不是JPA Entity；nullable逐项沿用Baseline 1.3，数据库generated活动键不公开。
...
    UserRoleResponse:
      type: object
      additionalProperties: false
      properties:
        userId:
          type: string
          pattern: ^[1-9][0-9]*$
          maxLength: 19
          description: SQL signed BIGINT正整数十进制字符串；服务端再校验<=9223372036854775807，避免JavaScript Number精度丢失。
          readOnly: true
        roleId:
          type: string
          pattern: ^[1-9][0-9]*$
          maxLength: 19
          description: SQL signed BIGINT正整数十进制字符串；服务端再校验<=9223372036854775807，避免JavaScript Number精度丢失。
          readOnly: true
        grantedBy:
          type: string
          pattern: ^[1-9][0-9]*$
          maxLength: 19
          description: SQL signed BIGINT正整数十进制字符串；服务端再校验<=9223372036854775807，避免JavaScript Number精度丢失。
          readOnly: true
        grantedAt:
          type: string
          format: date-time
          pattern: .*Z$
          description: 服务端UTC，ISO8601 Z；SQL DATETIME(6)，微秒精度。
          readOnly: true
      required:
      - userId
      - roleId
      - grantedBy
      - grantedAt
      description: 安全传输投影；不是JPA Entity；nullable逐项沿用Baseline 1.3，数据库generated活动键不公开。
...
    AdminUserResponse:
      type: object
      additionalProperties: false
      properties:
        userId:
          type: string
          pattern: ^[1-9][0-9]*$
          maxLength: 19
          description: SQL signed BIGINT正整数十进制字符串；服务端再校验<=9223372036854775807，避免JavaScript Number精度丢失。
          readOnly: true
        username:
          type: string
          maxLength: 64
          readOnly: true
        email:
          type: string
          maxLength: 254
          readOnly: true
        displayName:
          type: string
          maxLength: 100
          readOnly: true
        accountStatus:
          type: string
          maxLength: 32
          enum:
          - ENABLED
          - DISABLED
          readOnly: true
        createdAt:
          type: string
          format: date-time
          pattern: .*Z$
          description: 服务端UTC，ISO8601 Z；SQL DATETIME(6)，微秒精度。
          readOnly: true
        updatedAt:
          type: string
          format: date-time
          pattern: .*Z$
          description: 服务端UTC，ISO8601 Z；SQL DATETIME(6)，微秒精度。
          readOnly: true
        lockVersion:
          type: string
          pattern: ^(0|[1-9][0-9]*)$
          maxLength: 19
          description: 原有signed BIGINT lockVersion非负十进制字符串；服务端校验<=9223372036854775807、比较并由@Version更新。
          readOnly: true
        roles:
          type: array
          items:
            $ref: '#/components/schemas/RoleCode'
          uniqueItems: true
          description: 启用账号至少一个角色；禁用账号按原BR允许为空，不能用响应schema增加更强业务规则。
        assignments:
          type: array
          items:
            $ref: '#/components/schemas/UserRoleResponse'
      required:
      - userId
      - username
      - email
      - displayName
      - accountStatus
      - createdAt
      - updatedAt
      - lockVersion
      - roles
      - assignments
      description: 仅ADMIN；绝不返回密码或哈希。
...
    CreateUserRequest:
      type: object
      additionalProperties: false
      properties:
        username:
          type: string
          maxLength: 64
          minLength: 1
          pattern: .*\S.*
        email:
          type: string
          maxLength: 254
          minLength: 1
          pattern: .*\S.*
          format: email
        displayName:
          type: string
          maxLength: 100
          minLength: 1
          pattern: .*\S.*
        initialPassword:
          type: string
          minLength: 1
          pattern: .*\S.*
          writeOnly: true
          x-rms-max-utf8-bytes: 72
          description: 原始Java String；不trim、不做NFC/NFKC或其他Unicode normalization。满足既有non-blank contract且按原值UTF-8编码的实际长度为1..72 bytes；超过72 bytes以400 INVALID_INPUT拒绝，不静默截断。与登录password使用同一服务端byte-limit validator；仅服务器内存处理，校验后BCrypt生成非空哈希，不保存/记录/返回原密码。maxLength不代表字节限制；扩展属性须由服务端显式实现。
        roles:
          type: array
          minItems: 1
          uniqueItems: true
          items:
            $ref: '#/components/schemas/RoleCode'
      required:
      - username
      - email
      - displayName
      - initialPassword
      - roles
    EditUserRequest:
      type: object
      additionalProperties: false
      properties:
        username:
          type: string
          maxLength: 64
          minLength: 1
          pattern: .*\S.*
        email:
          type: string
          maxLength: 254
          minLength: 1
          pattern: .*\S.*
          format: email
        displayName:
          type: string
          maxLength: 100
          minLength: 1
          pattern: .*\S.*
        expectedLockVersion:
          $ref: '#/components/schemas/LockVersion'
      required:
      - expectedLockVersion
      description: 至少包含一个基本资料字段；不含密码、角色或状态。
      anyOf:
      - required:
        - username
      - required:
        - email
      - required:
        - displayName
    SetRolesRequest:
      type: object
      additionalProperties: false
      properties:
        roles:
          type: array
          items:
            $ref: '#/components/schemas/RoleCode'
          uniqueItems: true
        expectedLockVersion:
          $ref: '#/components/schemas/LockVersion'
      required:
      - roles
      - expectedLockVersion
      description: ENABLED账号至少保留一个角色；DISABLED是否无角色由原BR-USER-005判定，不增加永久至少一角色规则。
    LockCommand:
      type: object
      additionalProperties: false
      properties:
        expectedLockVersion:
          $ref: '#/components/schemas/LockVersion'
      required:
      - expectedLockVersion
...
    ApiError:
      type: object
      additionalProperties: false
      properties:
        code:
          type: string
          enum:
          - INVALID_INPUT
          - INCOMPLETE_CONTENT
          - UNAUTHENTICATED
          - INVALID_CREDENTIALS
          - FORBIDDEN
          - ACCOUNT_DISABLED
          - SELF_REVIEW
          - NOT_AUTHOR
          - NOT_FOUND
          - LOCK_VERSION_CONFLICT
          - STATE_CONFLICT
          - BASE_VERSION_STALE
          - DUPLICATE
          - ACTIVE_CHANGE_EXISTS
          - ACTIVE_REVIEW_EXISTS
          - DEPENDENCY_CYCLE
          - GRAPH_BUSY
          - INTERNAL_ERROR
        message:
          type: string
        correlationId:
          type: string
        currentLockVersion:
          $ref: '#/components/schemas/LockVersion'
      required:
      - code
      - message
      - correlationId
      description: 安全错误，无SQL/stack/密码；currentLockVersion仅有明确受权资源的修订冲突才提供。
    AdminUserResponsePage:
      type: object
      additionalProperties: false
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/AdminUserResponse'
        page:
          type: integer
          minimum: 0
        size:
          type: integer
          minimum: 1
          maximum: 100
        totalElements:
          type: integer
          minimum: 0
      required:
      - items
      - page
      - size
      - totalElements
...
    RequirementVersionResponsePage:
      type: object
      additionalProperties: false
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/RequirementVersionResponse'
        page:
          type: integer
          minimum: 0
        size:
          type: integer
          minimum: 1
          maximum: 100
        totalElements:
          type: integer
          minimum: 0
      required:
      - items
      - page
      - size
      - totalElements
...